backend/requirements.txt
fastapi>=0.110
uvicorn>=0.29
pydantic>=2.0
pytest>=8.0
httpx>=0.26
openai>=1.0.0
python-dotenv>=1.0